requirements.txt
numpy==2.2.6
torch==2.8.0
torchvision==0.23.0
opencv-python==4.12.0.88
pillow==11.3.0
scikit-image==0.25.2
scikit-learn==1.7.2
PyWavelets==1.8.0
h5py==3.14.0
pytorch-wavelets==1.3.0
